# server/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "AnimeDiary",
    "description": "API Specifications for Anime Diary project."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
  "paths": {
    "/api/login": {
      "post": {
        "summary": "Login to the system.",
        "tags": [
          "Session"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged into the system.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginSuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/register": {
      "post": {
        "summary": "Create a new account.",
        "tags": [
          "Session"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A new account is successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Unable to create an account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/animes": {
      "get": {
        "summary": "Search animes.",
        "tags": [
          "Animes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "path",
            "description": "Used for searching animes by. Searches all animes if empty value is passed.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "path",
            "description": "If passed, searches animes belong to the logged on user. Comma separated for multiple.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "path",
            "description": "The max number of animes to return. Default value of 10 is used when empty value is passed.",
            "required": true,
            "allowEmptyValue": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of animes returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimeListSuccessResponse"
                }
              }
            }
          },
          "500": {
            "description": "Somthing went wrong while searching.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a user's anime.",
        "tags": [
          "Animes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "animeId",
            "in": "path",
            "description": "The anime to update for the logged in user.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAnimeBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated anime returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimeSuccessResponse"
                }
              }
            }
          },
          "404": {
            "description": "Anime not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureResponse"
                }
              }
            }
          },
          "500": {
            "description": "Somthing went wrong while searching.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureResponse"
                }
              }
            }
          }
        }
      }
    }
  },

  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Header name: \"Authorization\", Sample Value: \"Bearer eyJhbGciOiJI...\""
      }
    },
    "schemas": {
      "LoginBody": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "userName",
          "password"
        ]
      },
      "RegisterBody": {
        "allOf": [
          {
            "$ref": "#/components/schemas/LoginBody"
          }
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "UpdateAnimeBody": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "number"
          },
          "notes": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "UserData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "UserDataWithToken": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UserData"
          }
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "AnimeData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "desc": {
            "type": "string"
          },
          "image": {
            "type": "string"
          }
        }
      },
      "UserAnimeData": {
        "type": "object",
        "properties": {
          "anime": {
            "$ref": "#/components/schemas/AnimeData"
          },
          "notes": {
            "type": "string"
          },
          "rating": {
            "type": "number"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AnimeSuccessResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/UserAnimeData"
          }
        }
      },
      "AnimeListSuccessResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "default": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserAnimeData"
            }
          }
        }
      },
      "RegisterSuccessResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/UserData"
          }
        }
      },
      "LoginSuccessResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/UserDataWithToken"
          }
        }
      },
      "FailureResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "default": false
          },
          "error": {
            "type": "string"
          }
        }
      }
    }
  }
}